+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Prevent add-to-cart to trigger view-cart, when view-cart is on cart-page only
+- Missing Quantity Parameter in add_to_cart Event
 
 ## [3.4.1] - 20 June 2023
 ### Fixed